Fill class-list from ROOTMAPs.
We get all rootmaps from the interpreter,
then filter them by user-provided regexps.
If the user provides nothing, we check
the lib-dir of ROOT and take all ROOTMAPs from there.
May later add an option to just take all ROOTMAPs
from everywhere.

Then, the ROOTMAPs are parsed (in a ROOT 5 and 6 compatible way),
and we test all contained classes.
This has lead to a large blow-up of the blacklist
(tested with ROOT 5 only for now) since many classes
fail a basic construction / destruction test.

We need to catch that somehow instead of segfaulting
to make things more convenient.