Reorganization to prepare for modular tests.
We now have:
- Headerfiles in include subdirectories.
- Tests in tests-subdirectory.
- A basic testInterface.
All tests should inherit from the testInterface,
and trigger static construction of themselves.
With that, they are registered in the global
list of tests via the testInterface.

Also, we have:
classObject - This is a TClass extended
with cached information which tests can use,
to check their applicability.
It also houses a list of already done tests.

All still work in progress, no functional changes
just yet.